refactor(ingest): split gitea.rs into gitea/{client,embed}.rs
src/ingest/gitea.rs was 510 lines, exceeding the 500-line monolith limit.
Split into three focused files:
- gitea.rs: module root with GiteaTarget, parse/normalize functions, ingest_gitea()
- gitea/client.rs: API types (GiteaRepo/User/Issue/Label/PullRequest), build_client(), fetch_repo(), fetch_paginated()
- gitea/embed.rs: all embedding helpers (payload, embed_metadata/issues/pulls, issue_doc, pull_doc)

Removes the .monolith-allowlist exemption entry (was: expires 2026-06-20).

Also fixes pre-existing clippy::unnecessary_qualification in subconfigs.rs
(std::fmt::Debug impl used std:: prefix despite `use std::fmt;` being present).

Verified: cargo test --lib -- gitea (5 passed), cargo test --lib -- ingest (225 passed),
cargo check --bin axon (exit 0), cargo clippy --lib (exit 0), monolith policy check passed.
