+++ src/ingest/gitea.rs
@@ -1,15 +1,17 @@
 use anyhow::{Result, anyhow, bail};
-use reqwest::header::{HeaderMap, HeaderValue};
-use reqwest::{Client, Url};
-use serde::Deserialize;
+use reqwest::Url;
 
 use crate::core::config::Config;
 use crate::core::http::validate_url;
 use crate::core::logging::{log_done, log_info, log_warn};
 use crate::ingest::generic_git::ingest_git_repository;
-use crate::ingest::git_payload::{GitPayload, build_git_payload};
 use crate::ingest::progress::PhaseReporter;
-use crate::vector::ops::{PreparedDoc, chunk_text, embed_prepared_docs};
+
+mod client;
+mod embed;
+
+use client::{build_client, fetch_repo};
+use embed::{embed_issues, embed_metadata, embed_pulls};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct GiteaTarget {
@@ -20,57 +22,17 @@ pub struct GiteaTarget {
     pub api_base: String,
 }
 
-#[derive(Debug, Deserialize)]
-struct GiteaRepo {
-    full_name: Option<String>,
-    description: Option<String>,
-    html_url: Option<String>,
-    clone_url: Option<String>,
-    default_branch: Option<String>,
-    private: Option<bool>,
-    stars_count: Option<u64>,
-    forks_count: Option<u64>,
-    open_issues_count: Option<u64>,
-}
+impl GiteaTarget {
+    pub(crate) fn as_normalized_target(&self) -> String {
+        format!("{}/{}/{}", self.host, self.owner, self.repo)
+    }
 
-#[derive(Debug, Deserialize)]
-struct GiteaUser {
-    login: Option<String>,
-    full_name: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct GiteaIssue {
-    number: u64,
-    title: String,
-    body: Option<String>,
-    state: Option<String>,
-    html_url: Option<String>,
-    user: Option<GiteaUser>,
-    labels: Option<Vec<GiteaLabel>>,
-    created_at: Option<String>,
-    updated_at: Option<String>,
-    comments: Option<u64>,
-}
-
-#[derive(Debug, Deserialize)]
-struct GiteaLabel {
-    name: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct GiteaPullRequest {
-    number: u64,
-    title: String,
-    body: Option<String>,
-    state: Option<String>,
-    html_url: Option<String>,
-    user: Option<GiteaUser>,
-    labels: Option<Vec<GiteaLabel>>,
-    created_at: Option<String>,
-    updated_at: Option<String>,
-    comments: Option<u64>,
-    merged: Option<bool>,
+    pub(crate) fn repo_api_url(&self, suffix: &str) -> String {
+        format!(
+            "{}/repos/{}/{}{}",
+            self.api_base, self.owner, self.repo, suffix
+        )
+    }
 }
 
 pub fn normalize_gitea_target(input: &str) -> Result<String> {
@@ -115,348 +77,6 @@ pub fn parse_gitea_target(input: &str) -> Result<GiteaTarget> {
         repo,
         web_url,
         api_base: format!("{}://{host}/api/v1", parsed.scheme()),
-    })
-}
-
-impl GiteaTarget {
-    fn as_normalized_target(&self) -> String {
-        format!("{}/{}/{}", self.host, self.owner, self.repo)
-    }
-
-    fn repo_api_url(&self, suffix: &str) -> String {
-        format!(
-            "{}/repos/{}/{}{}",
-            self.api_base, self.owner, self.repo, suffix
-        )
-    }
-}
-
-fn build_client(cfg: &Config) -> Result<Client> {
-    let mut headers = HeaderMap::new();
-    if let Some(token) = cfg.gitea_token.as_deref().filter(|token| !token.is_empty()) {
-        headers.insert(
-            "Authorization",
-            HeaderValue::from_str(&format!("token {token}"))?,
-        );
-    }
-    Ok(Client::builder()
-        .default_headers(headers)
-        .timeout(std::time::Duration::from_secs(60))
-        .build()?)
-}
-
-async fn fetch_repo(client: &Client, target: &GiteaTarget) -> Result<GiteaRepo> {
-    Ok(client
-        .get(target.repo_api_url(""))
-        .send()
-        .await?
-        .error_for_status()?
-        .json()
-        .await?)
-}
-
-async fn fetch_paginated<T: for<'de> Deserialize<'de>>(
-    client: &Client,
-    url: &str,
-    query: &[(&str, &str)],
-    max_items: usize,
-) -> Result<Vec<T>> {
-    let mut out = Vec::new();
-    let mut page = 1usize;
-    loop {
-        let mut request_url = Url::parse(url)?;
-        {
-            let mut pairs = request_url.query_pairs_mut();
-            for (key, value) in query {
-                pairs.append_pair(key, value);
-            }
-            pairs.append_pair("limit", "100");
-            pairs.append_pair("page", &page.to_string());
-        }
-        let mut page_items: Vec<T> = client
-            .get(request_url)
-            .send()
-            .await?
-            .error_for_status()?
-            .json()
-            .await?;
-        let page_len = page_items.len();
-        if max_items > 0 {
-            let remaining = max_items.saturating_sub(out.len());
-            page_items.truncate(remaining);
-        }
-        out.append(&mut page_items);
-        if page_len < 100 || (max_items > 0 && out.len() >= max_items) {
-            break;
-        }
-        page += 1;
-    }
-    Ok(out)
-}
-
-async fn embed_docs(cfg: &Config, docs: Vec<PreparedDoc>) -> Result<usize> {
-    let summary = embed_prepared_docs(cfg, docs, None)
-        .await
-        .map_err(|e| anyhow!("{e}"))?;
-    Ok(summary.chunks_embedded)
-}
-
-fn payload(
-    target: &GiteaTarget,
-    repo: &GiteaRepo,
-    kind: &'static str,
-    kind_extra: serde_json::Value,
-) -> serde_json::Value {
-    // Normalise Gitea "pull_request" to canonical "pr".
-    let git_content_kind: &'static str = if kind == "pull_request" { "pr" } else { kind };
-
-    let (state, number, author, labels, is_draft, merged_at, created_at, updated_at) = match kind {
-        "issue" | "pull_request" => (
-            kind_extra
-                .get("state")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-            kind_extra.get("number").and_then(|v| v.as_u64()),
-            kind_extra
-                .get("author")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-            kind_extra
-                .get("labels")
-                .and_then(|v| v.as_array())
-                .map(|a| {
-                    a.iter()
-                        .filter_map(|s| s.as_str().map(str::to_string))
-                        .collect()
-                }),
-            None::<bool>,
-            None::<String>,
-            kind_extra
-                .get("created_at")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-            kind_extra
-                .get("updated_at")
-                .and_then(|v| v.as_str())
-                .map(str::to_string),
-        ),
-        _ => (None, None, None, None, None, None, None, None),
-    };
-
-    build_git_payload(&GitPayload {
-        provider: "gitea".to_string(),
-        host: target.host.clone(),
-        owner: Some(target.owner.clone()),
-        repo: target.repo.clone(),
-        content_kind: git_content_kind,
-        branch: repo.default_branch.clone(),
-        state,
-        number,
-        author,
-        labels: labels.unwrap_or_default(),
-        is_draft,
-        merged_at,
-        created_at,
-        updated_at,
-        file_path: None,
-        file_language: None,
-        meta: Some(serde_json::json!({
-            "default_branch": repo.default_branch,
-            "private": repo.private,
-            "gitea": kind_extra,
-        })),
-    })
-}
-
-async fn embed_metadata(cfg: &Config, target: &GiteaTarget, repo: &GiteaRepo) -> Result<usize> {
-    let mut parts = Vec::new();
-    if let Some(desc) = repo.description.as_deref().filter(|desc| !desc.is_empty()) {
-        parts.push(format!("Description: {desc}"));
-    }
-    if let Some(stars) = repo.stars_count {
-        parts.push(format!("Stars: {stars}"));
-    }
-    if let Some(forks) = repo.forks_count {
-        parts.push(format!("Forks: {forks}"));
-    }
-    if parts.is_empty() {
-        return Ok(0);
-    }
-    let title = repo
-        .full_name
-        .clone()
-        .unwrap_or_else(|| format!("{}/{}", target.owner, target.repo));
-    let chunks = chunk_text(&format!("# {title}\n\n{}", parts.join("\n")));
-    if chunks.is_empty() {
-        return Ok(0);
-    }
-    embed_docs(
-        cfg,
-        vec![PreparedDoc {
-            url: repo
-                .html_url
-                .clone()
-                .unwrap_or_else(|| target.web_url.clone()),
-            domain: target.host.clone(),
-            chunks,
-            source_type: "gitea".to_string(),
-            content_type: "text",
-            title: Some(title),
-            extra: Some(payload(
-                target,
-                repo,
-                "repo_metadata",
-                serde_json::json!({
-                    "private": repo.private,
-                    "stars": repo.stars_count,
-                    "forks": repo.forks_count,
-                    "open_issues": repo.open_issues_count,
-                }),
-            )),
-            extractor_name: None,
-            structured: None,
-        }],
-    )
-    .await
-}
-
-fn author_name(user: &Option<GiteaUser>) -> Option<String> {
-    user.as_ref()
-        .and_then(|u| u.login.clone().or_else(|| u.full_name.clone()))
-}
-
-fn label_names(labels: Option<Vec<GiteaLabel>>) -> Vec<String> {
-    labels
-        .unwrap_or_default()
-        .into_iter()
-        .map(|l| l.name)
-        .collect()
-}
-
-async fn embed_issues(
-    cfg: &Config,
-    client: &Client,
-    target: &GiteaTarget,
-    repo: &GiteaRepo,
-) -> Result<usize> {
-    let issues: Vec<GiteaIssue> = fetch_paginated(
-        client,
-        &target.repo_api_url("/issues"),
-        &[("state", "all"), ("type", "issues")],
-        cfg.github_max_issues,
-    )
-    .await?;
-    let docs = issues
-        .into_iter()
-        .filter_map(|i| issue_doc(target, repo, i))
-        .collect();
-    embed_docs(cfg, docs).await
-}
-
-fn issue_doc(target: &GiteaTarget, repo: &GiteaRepo, issue: GiteaIssue) -> Option<PreparedDoc> {
-    let labels = label_names(issue.labels);
-    let label_text = if labels.is_empty() {
-        String::new()
-    } else {
-        format!("\nLabels: {}", labels.join(", "))
-    };
-    let content = format!(
-        "# Issue #{}: {}\n\n{}{}",
-        issue.number,
-        issue.title,
-        issue.body.as_deref().unwrap_or(""),
-        label_text
-    );
-    let chunks = chunk_text(&content);
-    if chunks.is_empty() {
-        return None;
-    }
-    Some(PreparedDoc {
-        url: issue
-            .html_url
-            .unwrap_or_else(|| format!("{}/issues/{}", target.web_url, issue.number)),
-        domain: target.host.clone(),
-        chunks,
-        source_type: "gitea".to_string(),
-        content_type: "text",
-        title: Some(format!("Issue #{}: {}", issue.number, issue.title)),
-        extra: Some(payload(
-            target,
-            repo,
-            "issue",
-            serde_json::json!({
-                "number": issue.number,
-                "state": issue.state,
-                "author": author_name(&issue.user),
-                "labels": labels,
-                "created_at": issue.created_at,
-                "updated_at": issue.updated_at,
-                "comment_count": issue.comments,
-            }),
-        )),
-        extractor_name: None,
-        structured: None,
-    })
-}
-
-async fn embed_pulls(
-    cfg: &Config,
-    client: &Client,
-    target: &GiteaTarget,
-    repo: &GiteaRepo,
-) -> Result<usize> {
-    let pulls: Vec<GiteaPullRequest> = fetch_paginated(
-        client,
-        &target.repo_api_url("/pulls"),
-        &[("state", "all")],
-        cfg.github_max_prs,
-    )
-    .await?;
-    let docs = pulls
-        .into_iter()
-        .filter_map(|p| pull_doc(target, repo, p))
-        .collect();
-    embed_docs(cfg, docs).await
-}
-
-fn pull_doc(target: &GiteaTarget, repo: &GiteaRepo, pull: GiteaPullRequest) -> Option<PreparedDoc> {
-    let labels = label_names(pull.labels);
-    let content = format!(
-        "# PR #{}: {}\n\n{}",
-        pull.number,
-        pull.title,
-        pull.body.as_deref().unwrap_or("")
-    );
-    let chunks = chunk_text(&content);
-    if chunks.is_empty() {
-        return None;
-    }
-    Some(PreparedDoc {
-        url: pull
-            .html_url
-            .unwrap_or_else(|| format!("{}/pulls/{}", target.web_url, pull.number)),
-        domain: target.host.clone(),
-        chunks,
-        source_type: "gitea".to_string(),
-        content_type: "text",
-        title: Some(format!("PR #{}: {}", pull.number, pull.title)),
-        extra: Some(payload(
-            target,
-            repo,
-            "pull_request",
-            serde_json::json!({
-                "number": pull.number,
-                "state": pull.state,
-                "author": author_name(&pull.user),
-                "labels": labels,
-                "created_at": pull.created_at,
-                "updated_at": pull.updated_at,
-                "comment_count": pull.comments,
-                "merged": pull.merged,
-            }),
-        )),
-        extractor_name: None,
-        structured: None,
     })
 }
